+++ support/resources/resources_windows.go
@@ -1,4 +1,6 @@
-package ek
+// Package resources provides methods for collecting information about system
+// resources (cpu/memory)
+package resources
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -7,5 +9,13 @@ package ek
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// VERSION is current ek package version
-const VERSION = "13.5.0"
+import (
+	"github.com/essentialkaos/ek/v13/support"
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// Collect collects info about CPU and memory
+func Collect() *support.ResourcesInfo {
+	return nil
+}